reputation fixes
The value of a government's attitude towards another government shouldn't be too large or too small. Because of this, the XSZ can be friendly even if you joined the Enclave. This commit fixes that. I also don't get why the XSZ would trust the player too much at the start of the story.